+++ u-sdk/src/oss/region.rs
@@ -1,7 +1,7 @@
 //! [API 文档](https://help.aliyun.com/zh/oss/developer-reference/describeregions)
 
 use super::Client;
-use super::utils::{get_request_header_with_bucket_region, parse_xml_response};
+use super::utils::{get_request_header, parse_xml_response};
 use crate::oss::Error;
 use crate::oss::sign_v4::HTTPVerb;
 use bon::Builder;
@@ -42,8 +42,10 @@ impl DescribeRegions<'_> {
         )
         .unwrap();
 
-        let header_map = get_request_header_with_bucket_region(
-            client,
+        let creds = client.credentials_provider.load().await?;
+        let header_map = get_request_header(
+            &creds.access_key_id,
+            &creds.access_key_secret,
             HashMap::with_capacity(0),
             &request_url,
             HTTPVerb::Get,